dont use deprecated casadi structs
