+++ examples/acados_python/pendulum_on_cart/ocp/ocp_example_cost_formulations.py
@@ -30,6 +30,7 @@
 
 import sys
 sys.path.insert(0, '../common')
+sys.path.insert(0, '../../chain_mass')
 
 from acados_template import AcadosOcp, AcadosOcpSolver, AcadosMultiphaseOcp
 from pendulum_model import export_pendulum_ode_model, export_augmented_pendulum_model
@@ -37,7 +38,7 @@ import numpy as np
 import scipy.linalg
 from utils import plot_pendulum
 import casadi as ca
-from casadi.tools import entry, struct_symSX
+from param_utils import ParamLayout, ParamVector
 
 COST_VERSIONS = ['LS', 'EXTERNAL', 'EXTERNAL_Z', 'NLS', 'NLS_TO_EXTERNAL', 'NLS_Z', 'LS_Z', 'CONL', 'CONL_Z', 'AUTO']
 HESSIAN_APPROXIMATION = 'GAUSS_NEWTON' # 'GAUSS_NEWTON
@@ -203,12 +204,14 @@ def formulate_ocp(cost_version: str, constraint_version="bu") -> AcadosOcp:
         ocp.cost.cost_type = 'NONLINEAR_LS'
         ocp.cost.cost_type_e = 'NONLINEAR_LS'
 
-        p_global = struct_symSX([
-            entry('W', shape=(ny, ny)),
-            entry('yref', shape=(ny, )),
-            entry('W_e', shape=(ny_e, ny_e)),
-            entry('yref_e', shape=(ny_e, ))
+        p_global_layout = ParamLayout([
+            {'name': 'W', 'shape': (ny, ny)},
+            {'name': 'yref', 'shape': (ny, )},
+            {'name': 'W_e', 'shape': (ny_e, ny_e)},
+            {'name': 'yref_e', 'shape': (ny_e, )}
         ])
+        p_global_sym = ca.SX.sym('p_global', p_global_layout.size, 1)
+        p_global = ParamVector(p_global_layout, p_global_sym)
         ocp.model.p_global = p_global.cat
 
         ocp.cost.W = p_global['W']
@@ -221,13 +224,13 @@ def formulate_ocp(cost_version: str, constraint_version="bu") -> AcadosOcp:
 
         ocp.translate_cost_to_external_cost(cost_hessian='GAUSS_NEWTON')
 
-        p_global_values = p_global(0)
+        p_global_values = ParamVector(p_global_layout, np.zeros((p_global_layout.size, 1)))
         p_global_values['W'] = cost_W
         p_global_values['yref'] = np.zeros((ny, ))
         p_global_values['W_e'] = Q_mat
         p_global_values['yref_e'] = np.zeros((ny_e, ))
 
-        ocp.p_global_values = p_global_values.cat.full().flatten()
+        ocp.p_global_values = p_global_values.cat.flatten()
     else:
         raise Exception('Unknown cost_version.')
